Moving Common/node to lib/.
