+++ docs/main.html
@@ -15,7 +15,7 @@ module.  The application is organized in the following manner:</p>
 
 <p>This is the sample implementation which deals with data for this application</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-2">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-2">&#182;</a>               </div>               <p>Include express and swagger in the application.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="kd">var</span> <span class="nx">express</span> <span class="o">=</span> <span class="nx">require</span><span class="p">(</span><span class="s2">&quot;express&quot;</span><span class="p">)</span>
  <span class="p">,</span> <span class="nx">url</span> <span class="o">=</span> <span class="nx">require</span><span class="p">(</span><span class="s2">&quot;url&quot;</span><span class="p">)</span>
- <span class="p">,</span> <span class="nx">swagger</span> <span class="o">=</span> <span class="nx">require</span><span class="p">(</span><span class="s2">&quot;../../Common/node/swagger.js&quot;</span><span class="p">);</span>
+ <span class="p">,</span> <span class="nx">swagger</span> <span class="o">=</span> <span class="nx">require</span><span class="p">(</span><span class="s2">&quot;../../lib/swagger.js&quot;</span><span class="p">);</span>
 
 <span class="kd">var</span> <span class="nx">petResources</span> <span class="o">=</span> <span class="nx">require</span><span class="p">(</span><span class="s2">&quot;./petResources.js&quot;</span><span class="p">);</span>
